refactor(combine refs): hoist common method for combining refs as a utility method
