+++ src/components/multi-select/multi-select.jsx
@@ -7,29 +7,8 @@ import React, {
 import PropTypes from 'prop-types';
 import FormControl from '../form-control/form-control.jsx';
 import MultiSelectControl from '../control/multi-select.jsx';
+import combineRefs from '../../utils/combine-refs.js';
 import styles from './multi-select.css';
-
-function spread(ref1, ref2) {
-  const newObject = {};
-
-  Object.keys(ref1.current).reduce((acc, key) => (
-    Object.defineProperty(acc, key, {
-      get() { return ref1.current[key]; },
-      configurable: true,
-      enumerable: true,
-    })
-  ), newObject);
-
-  Object.keys(ref2.current).reduce((acc, key) => (
-    Object.defineProperty(acc, key, {
-      get() { return ref2.current[key]; },
-      configurable: true,
-      enumerable: true,
-    })
-  ), newObject);
-
-  return newObject;
-}
 
 const MultiSelect = forwardRef(function MultiSelect(props, ref) {
   const [validationMessage, setValidationMessage] = useState(props.validationMessage);
@@ -51,7 +30,7 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
     setValidationMessage(event.detail.validationMessage);
   }
 
-  useImperativeHandle(ref, () => spread(
+  useImperativeHandle(ref, () => combineRefs(
     refs.control,
     refs.multiSelect,
   ));
